REF: share searchsorted between DTI/TDI/PI, insert between DTI/TDI
